+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# TBD - TBD
+- Fixed the Virtual Judge parser
+
 # 2.12.0 - January 31st, 2020
 - Added a parser for NowCoder (thanks [@landcold7](https://github.com/landcold7))
 - Added support for CP Editor